# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: getCustomers
      tags: [customers]
      summary: List all customers
      responses:
        "200":
          description: List of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
    post:
      operationId: createCustomer
      tags: [customers]
      summary: Create a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerInput"
      responses:
        "201":
          description: Customer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict - duplicate field
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customers/{id}:
    put:
      operationId: updateCustomer
      tags: [customers]
      summary: Update a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerInput"
      responses:
        "200":
          description: Customer updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteCustomer
      tags: [customers]
      summary: Delete a customer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /customer-quotations:
    get:
      operationId: getCustomerQuotations
      tags: [customer-quotations]
      summary: List all customer quotations
      responses:
        "200":
          description: List of quotations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CustomerQuotation"
    post:
      operationId: createCustomerQuotation
      tags: [customer-quotations]
      summary: Create a new customer quotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerQuotationInput"
      responses:
        "201":
          description: Quotation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerQuotation"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /customer-quotations/{id}:
    get:
      operationId: getCustomerQuotation
      tags: [customer-quotations]
      summary: Get a single customer quotation with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quotation with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomerQuotation"
    delete:
      operationId: deleteCustomerQuotation
      tags: [customer-quotations]
      summary: Delete a customer quotation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
components:
  schemas:
...
    Customer:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        status:
          type: string
        createdAt:
          type: string
      required:
        - id
        - name
        - status
        - createdAt
    CreateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
      required:
        - name
    UpdateCustomerInput:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        address:
          type: string
        commercialReg:
          type: string
        taxReg:
          type: string
        status:
          type: string
      required:
        - name
    CustomerQuotationItem:
      type: object
      properties:
        id:
          type: integer
        quotationId:
          type: integer
        customerItemCode:
          type: string
        description:
          type: string
        partNo:
          type: string
        unit:
          type: string
        quantity:
          type: string
        sortOrder:
          type: integer
      required:
        - id
        - quotationId
        - description
        - quantity
    CustomerQuotation:
      type: object
      properties:
        id:
          type: integer
        quotationNo:
          type: string
        customerId:
          type: integer
        customerName:
          type: string
        responsibleName:
          type: string
        requestDate:
          type: string
        expiryDate:
          type: string
        customerOrderNo:
          type: string
        status:
          type: string
        createdAt:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CustomerQuotationItem"
      required:
        - id
        - quotationNo
        - customerId
        - requestDate
        - status
        - createdAt
    CreateCustomerQuotationInput:
      type: object
      properties:
        customerId:
          type: integer
        responsibleName:
          type: string
        requestDate:
          type: string
        expiryDate:
          type: string
        customerOrderNo:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              customerItemCode:
                type: string
              description:
                type: string
              partNo:
                type: string
              unit:
                type: string
              quantity:
                type: number
            required:
              - description
              - quantity
      required:
        - customerId
        - requestDate
        - items
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success